DEMOS-1709: edit and delete state files by SU
Before submitted.

+++ server/src/model/document/handleDeleteDocument.ts
@@ -1,4 +1,5 @@
 import { Document as PrismaDocument } from "@prisma/client";
+import type { ContextUser } from "../../auth";
 import { PrismaTransactionClient } from "../../prismaClient";
 import { S3Adapter } from "../../adapters/s3/S3Adapter";
 import { deleteDocumentById } from ".";
@@ -7,10 +8,20 @@ import { validateDocumentCanBeDeleted } from "./validateDocumentCanBeDeleted";
 export async function handleDeleteDocument(
   tx: PrismaTransactionClient,
   s3Adapter: S3Adapter,
-  id: string
+  id: string,
+  user?: ContextUser
 ): Promise<PrismaDocument> {
+  const documentToDelete = await tx.document.findUnique({
+    where: { id },
+    include: { deliverable: { select: { statusId: true } } },
+  });
+
+  if (!documentToDelete) {
+    throw new Error(`Document with ID ${id} was not found.`);
+  }
+
+  validateDocumentCanBeDeleted(documentToDelete, user);
   const document = await deleteDocumentById(tx, id);
-  validateDocumentCanBeDeleted(document);
   const key = `${document.applicationId}/${document.id}`;
   await s3Adapter.moveDocumentFromCleanToDeleted(key);
   return document;
